pyproject.toml
[project]
name = "noctua"
version = "0.1.0"
description = "AI-native CLI for Obsidian.md to superpower your learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chonkie[model2vec]>=1.5.4",
    "markitdown[pdf]>=0.1.4",
    "mcp>=1.0.0",
    "openrouter>=0.6.0",
    "pydantic-ai>=1.53.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "questionary>=2.1.1",
    "rich>=14.3.2",
    "sentence-transformers>=5.2.2",
    "txtai>=9.4.1",
    "typer>=0.21.1",
    "pydantic-ai-slim[duckduckgo]>=1.53.0",
    "uvicorn>=0.40.0",
    "fastapi>=0.134.0",
    "scholarly>=1.7.11",
]

[project.scripts]
noctua = "noctua.cli.main:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["noctua*"]

[tool.uv]
package = true
